Create Cred Exil on Screenlock
This Bash script is designed to work with Hak5's KeyCroc for advanced credential capture and payload delivery on Windows 8/10/11 systems. The script listens for user input to capture username and password details, waits for a lock screen event (Win + L), and automatically logs back into the system upon detecting an unlock attempt. Once logged in, it delivers a payload to extract a wide range of sensitive data